Add 'components' directory and move the toolbar provider definition to a separate file.
